Remove the local sprinkles
